+++ src/foreign_link.rs
@@ -1,4 +1,4 @@
-use crate::{DatabaseAccess, DatabaseRecord, Record, ServiceError};
+use crate::{DatabaseAccess, DatabaseRecord, Error, Record};
 
 /// The `ForeignLink` trait of the Aragog library.
 /// It allows to define foreign_key relations between different models.
@@ -77,7 +77,7 @@ pub trait ForeignLink<T: Record>: Sized {
 
     /// Retrieves the record matching the defined `foreign_key`. Type inference may be required.
     #[cfg(feature = "async")]
-    async fn linked_model<D>(&self, db_access: &D) -> Result<DatabaseRecord<T>, ServiceError>
+    async fn linked_model<D>(&self, db_access: &D) -> Result<DatabaseRecord<T>, Error>
     where
         Self: Sized,
         T: 'async_trait,
@@ -88,7 +88,7 @@ pub trait ForeignLink<T: Record>: Sized {
 
     /// Retrieves the record matching the defined `foreign_key`. Type inference may be required.
     #[cfg(not(feature = "async"))]
-    fn linked_model<D>(&self, db_access: &D) -> Result<DatabaseRecord<T>, ServiceError>
+    fn linked_model<D>(&self, db_access: &D) -> Result<DatabaseRecord<T>, Error>
     where
         D: DatabaseAccess + ?Sized,
     {