+++ src/update.rs
@@ -1,4 +1,4 @@
-use crate::ServiceError;
+use crate::Error;
 
 /// The `Update` trait of the Aragog library.
 /// This trait provides the possibility to update a Type from an other one. Its main use
@@ -8,12 +8,12 @@ use crate::ServiceError;
 pub trait Update<T> {
     /// Update the `Self` field values `T`. The object takes a mutable reference of itself and is directly
     /// updated.
-    /// Can fail and return an error, the error is in most of the cases a [`ServiceError`]::[`ValidationError`]
+    /// Can fail and return an error, the error is in most of the cases a [`Error`]::[`ValidationError`]
     /// on fields validation failure
     ///
-    /// [`ServiceError`]: enum.ServiceError.html
-    /// [`ValidationError`]: enum.ServiceError.html#variant.ValidationError
-    fn update(&mut self, form: &T) -> Result<(), ServiceError>;
+    /// [`Error`]: enum.Error.html
+    /// [`ValidationError`]: enum.Error.html#variant.ValidationError
+    fn update(&mut self, form: &T) -> Result<(), Error>;
 
     /// Can update a mutable `value` with a new one if the `new_value` is defined (`Some`).
     /// if the `new_value` is `None` the value stays unchanged